Moved layout elements to own `layout` folder
