+++ frontend/src/components/layout/ContentCard.vue
