+++ frontend/src/components/layout/ContentGroup.vue
