+++ frontend/src/views/Home.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script>
-import ContentCard from '@/components/base/ContentCard'
+import ContentCard from '@/components/layout/ContentCard'
 
 export default {
   name: 'Home',
